% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fans.R
\name{centile_fan_lifespan}
\alias{centile_fan_lifespan}
\title{Plot lifespan centile fan}
\usage{
centile_fan_lifespan(
  gamlssModel,
  df,
  x_var = "logAge",
  color_var = "sex",
  desiredCentiles = c(0.025, 0.5, 0.975),
  sim_grid_list = NULL,
  ...
)
}
\value{
ggplot object
}
\description{
Plot centiles in the style of Bethlehem, Seidletz & White et al, Nature 2020.
}
\details{
Wrapper function for \code{\link[=make_centile_fan]{make_centile_fan()}} with different defaults.
}
