% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{cohens_f2_local}
\alias{cohens_f2_local}
\title{Cohen's Fsquared Local}
\usage{
cohens_f2_local(full_mod, null_mod)
}
\arguments{
\item{full_mod}{full gamlss model object}

\item{null_mod}{null gamlss model object (refit without covariate of interest)}
}
\value{
numeric fsquared value
}
\description{
Calculate effect size (cohen's fsq) of a covariate using the difference in Rsq of full and nested models.
}
\details{
See Equation 2 in \href{https://www.frontiersin.org/journals/psychology/articles/10.3389/fpsyg.2012.00111/full}{Selya et al, 2012}
}
\section{Optional dependency}{

Requires the suggested package gamlss2, whose \code{Rsq()} has methods for both
gamlss and gamlss2 fits. See \link{gamlssTools-optional}.
}

\examples{
iris_model_full <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris)
iris_model_null <- gamlss(formula = Sepal.Width ~ Sepal.Length, sigma.formula = ~ Sepal.Length, data=iris)

cohens_f2_local(iris_model_full, iris_model_null)

}
