% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{gamlss_ci}
\alias{gamlss_ci}
\title{Bootstrap Confidence Intervals}
\usage{
gamlss_ci(
  boot_list,
  x_var,
  factor_var = NULL,
  special_term = NULL,
  moment = c("mu", "sigma"),
  interval = 0.95,
  sliding_window = FALSE,
  df
)
}
\arguments{
\item{boot_list}{output of gamlssTools::bootstrap_gamlss()}

\item{x_var}{numeric variable to plot confidence intervals across}

\item{factor_var}{categorical variable, CIs will be calculated separately at each level.}

\item{special_term}{optional, passed to gamlssTools::sim_data()}

\item{moment}{what moment to get CIs for. Currently only implemented for mu, which returns 50th percentile CIs}

\item{interval}{size of confidence interval to calculate. Defaults to 0.95, or 95\%}

\item{sliding_window}{logical indicating whether to calculate CI at 500 point clusters along x or use sliding windows.
Defaults to FALSE}

\item{df}{true dataframe}
}
\value{
list of dataframes, with one dataframe for each level of \code{factor_var}
}
\description{
Uses bootstrapped gamlss models to calculate CIs
}
\details{
Takes output of gamlssTools::bootstrap_gamlss() and uses them to calculate confidence intervals for
the 50th percentiles across the range of \code{x_var}.
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris)
boot_out <- bootstrap_gamlss(iris_model, df=iris, type="resample", stratify=TRUE, group_var="Species")
gamlss_ci(boot_out, "Sepal.Length", "Species", df=iris)

}
