% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{get.mu.coeff}
\alias{get.mu.coeff}
\title{Get Mu Beta}
\usage{
get.mu.coeff(gamlssModel, term)
}
\arguments{
\item{gamlssModel}{gamlss model object}

\item{term}{coefficient to return beta of}
}
\value{
beta weight for given \code{term} in mu (numeric)
}
\description{
Extract beta weight of a term in a gamlss model's mu parameter
}
\details{
Only works for fixed effects (not random effects)
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris)
get.mu.coeff(iris_model, "Sepal.Length")

}
