% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{get_coeff}
\alias{get_coeff}
\title{Get Coefficient}
\usage{
get_coeff(gamlssModel, moment, term)
}
\arguments{
\item{gamlssModel}{gamlss model object}

\item{moment}{moment containing \code{term}}

\item{term}{coefficient to return beta of}
}
\value{
beta weight for given \code{term} in \code{moment} (numeric)
}
\description{
Extract beta weight of a term in a gamlss model
}
\details{
Only works for fixed effects (not random effects)
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris)
get_coeff(iris_model, "mu", "Sepal.Length")

}
