% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{get_y}
\alias{get_y}
\title{Get y}
\usage{
get_y(gamlssModel)
}
\arguments{
\item{gamlssModel}{gamlss model object}
}
\value{
y variable name
}
\description{
Silly little function to extract y from gamlss model
}
