% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{GGalt.variance}
\alias{GGalt.variance}
\title{GG variance}
\usage{
GGalt.variance(mu, sigma, nu)
}
\arguments{
\item{mu}{vector of mu parameter predicted values}

\item{sigma}{vector of mu parameter predicted values}

\item{nu}{vector of mu parameter predicted values}
}
\value{
numeric vector of variance values
}
\description{
Extract variance of gamlss model with GG distribution
}
\details{
Written by Simon White and copied from \href{https://github.com/brainchart/Lifespan/blob/bca92bfaad13cada8aad60cd14bc0bdaeb194ad7/102.gamlss-recode.r#L90}{github}
}
