% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fan_helper_funs.R
\name{med_diff}
\alias{med_diff}
\title{Get Median Diffs across X}
\usage{
med_diff(gamlssModel, df, x_var, factor_var, sim_data_list = NULL, ...)
}
\arguments{
\item{gamlssModel}{gamlss model object}

\item{df}{dataframe model was originally fit on}

\item{x_var}{continuous predictor (e.g. 'age'), which \code{sim_df_list} varies over#'}

\item{factor_var}{categorical variable to compare levels within.}

\item{sim_df_list}{list of simulated dataframes returned by \code{sim_data()}}
}
\value{
dataframe
}
\description{
Calculate differences in 50th centile trajectories between 2 factor levels
}
\details{
To test significance, see gamlssTools::ci_diffs()
}
