% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{peak_ci}
\alias{peak_ci}
\title{Bootstrap CIs for Peak Values}
\usage{
peak_ci(
  boot_list,
  x_var,
  factor_var = NULL,
  special_term = NULL,
  moment = c("mu", "sigma"),
  interval = 0.95,
  df = NULL,
  sim_data_list = NULL
)
}
\arguments{
\item{boot_list}{output of gamlssTools::bootstrap_gamlss()}

\item{x_var}{numeric variable to plot confidence intervals across}

\item{factor_var}{categorical variable, CIs will be calculated separately at each level.}

\item{special_term}{optional, passed to gamlssTools::sim_data()}

\item{moment}{what moment to get CIs for. Currently only implemented for mu, which returns 50th percentile CIs}

\item{interval}{size of confidence interval to calculate. Defaults to 0.95, or 95\%}

\item{df}{true dataframe (optional, must pass this or \code{sim_data_list})}

\item{sim_data_list}{data simulated from true dataframe (optional, must pass this or \code{df})}
}
\value{
list of dataframes, with one dataframe for each level of \code{factor_var}
}
\description{
Calculate CIs around age at peak
}
\details{
Takes output of gamlssTools::bootstrap_gamlss() and uses them to calculate confidence intervals for
the value of \code{x_var} where the 50th precentile value peaks
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris)
boot_out <- bootstrap_gamlss(iris_model, df=iris, type="resample", stratify=TRUE, group_var="Species")
peak_ci(boot_out, "Sepal.Length", "Species", df=iris)

}
