% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fans.R
\name{plot_centile_cis}
\alias{plot_centile_cis}
\title{Plot with CIs}
\usage{
plot_centile_cis(
  gamlssModel,
  df,
  x_var,
  color_var,
  desiredCentiles = c(0.5),
  interval = 0.95,
  B = 100,
  sim_data_list = NULL,
  type = c("resample", "bayes", "LOSO"),
  stratify = FALSE,
  boot_group_var = NULL,
  special_term = NULL,
  ...
)
}
\value{
ggplot object
}
\description{
Plot centile fan with confidence intervals on 50th centile
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris, family=BCCG)
plot_centile_cis(iris_model, iris, "Sepal.Length", "Species", stratify=TRUE, boot_group_var="Species")

}
