% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{plot_sigma_cis}
\alias{plot_sigma_cis}
\title{Plot sigma with CIs}
\usage{
plot_sigma_cis(
  gamlssModel,
  df,
  x_var,
  color_var,
  interval = 0.95,
  B = 100,
  sim_data_list = NULL,
  type = c("resample", "bayes", "LOSO"),
  stratify = FALSE,
  boot_group_var = NULL,
  special_term = NULL,
  boot_list = NULL,
  average_over = FALSE,
  ci_type = c("pointwise", "sliding", "simultaneous"),
  ...
)
}
\arguments{
\item{gamlssModel}{gamlss model object}

\item{df}{dataframe used to fit the gamlss model}

\item{x_var}{continuous predictor (e.g. 'age') that will be plotted on the x axis}

\item{color_var}{(optional) categorical predictor (e.g. 'sex') that will be used to determine the color of
points/centile lines. Alternatively, you can average over each level of this variable
to return a single set of centile lines (see \code{average_over}).}

\item{interval}{size of confidence interval to calculate. Defaults to 0.95, or 95\%}

\item{B}{(optional) number of samples/models to bootstrap. Defaults to 100. if \code{type = "LOSO"}, B will be updated to
the number of unique values of \code{group_var}}

\item{sim_data_list}{(optional) output of \code{sim_data()}.}

\item{type}{(optional) which type of bootstrapping to perform. \code{resample} performs traditional bootstrapping (resample with replacement)
across all groups; alternatively, it may be combined with \code{stratify=TRUE} and \code{group_var} args below to bootstrap
while maintaining each group's (e.g study's) n. \code{bayes} keeps the original dataframe but randomizes each observation's
weight. \code{LOSO} drops an entire subset from the sample (indicated by \code{group_var}) with each bootstrap.}

\item{stratify}{(optional) logical. with \code{type=resample} will bootstrap within each level of \code{group_var}.}

\item{boot_group_var}{(optional) categorical/factor variable that resampling will be stratified within (when \code{type=resample})
or that one level will be dropped from in each bootstraped sample (when \code{type=LOSO}). Can also be a list, allowing
stratification within multiple groups e.g. \code{group_var=c(sex, study)}}

\item{special_term}{(optional) passed to \code{sim_data()}}

\item{boot_list}{(optional) output of \code{bootstrap_gamlss()}}

\item{average_over}{logical indicating whether to average predicted centiles across each level of \code{color_var}.}

\item{ci_type}{options for type of precentile CI to return. \code{pointwise} (default) calculates percentiles at 500 points
along \code{x_var}. \code{sliding} does the same with a sliding window. \code{simultaneous} implements simultaneous CIs along \code{x_var}
as described in Gao et al (doi: 10.3390/sym13071212).}
}
\value{
ggplot object
}
\description{
Plot sigma with confidence intervals
}
\details{
Wrapper function for \code{plot_sigma()} and \code{gamlss_ci()}
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris, family=BCCG)
plot_sigma_cis(iris_model, iris, "Sepal.Length", "Species", stratify=TRUE, boot_group_var="Species")

}
