% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{pointwise_pct_ci}
\alias{pointwise_pct_ci}
\title{Pointwise Confidence Intervals}
\usage{
pointwise_pct_ci(df, x_var, col_name, interval)
}
\description{
Helper function for \code{gamlss_ci()}, \code{ci_type} = "pointwise" (default)
}
