% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fan_helper_funs.R
\name{pred_residualized}
\alias{pred_residualized}
\title{Predict residualized response}
\usage{
pred_residualized(gamlssModel, new_data, og_data = NULL, rm_terms)
}
\arguments{
\item{gamlssModel}{gamlss model object}

\item{new_data}{dataframe to predict from. may be simulated data created with \code{\link[=sim_data]{sim_data()}}}

\item{og_data}{original dataframe used to fit the model (optional)}

\item{rm_terms}{list of term(s) whose effects will be residualized (removed).}
}
\value{
list of predicted values, corrected for effects listed in \code{rm_terms}.
}
\description{
Correct predicted response to remove the effects of certain unwanted covariates from simulated data
}
\details{
Works by predicting moment on link scale, subtracting beta weights for specified terms (\code{rm_terms}) as estimated by gamlss model
from all moments, then applying the inverse of the link function to convert into response scale.
Resulting responses can be fed into \code{\link[=pred_centile]{pred_centile()}} to calculate centiles from moments calculated WITHOUT
unwanted effects (such as site or study effects). Works for random effects but not other smooths.

IMPORTANT: assumes that data is simulated so that the values of \code{rm_terms} in \code{new_data} are constant.
If you want to residualize covariate effects from real/variable data, use \code{\link[=resid_data]{resid_data()}}.
}
\examples{
#predict a specific centile value across simulated data
iris_model <- gamlss(formula = Sepal.Width ~ fp(Sepal.Length) + Petal.Width + re(random=~1|Species), 
  sigma.formula = ~ Sepal.Length + random(Species), data=iris, family=BCCG)
sim_df <- sim_data(iris, "Sepal.Length", gamlssModel=iris_model)
pred_df <- pred_residualized(iris_model, new_data = sim_df[[1]], og_data=iris, 
  rm_terms=c("Species", "Petal.Width"))
pred_centile(0.1, pred_df, "qBCCG", 3)

}
