% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fan_helper_funs.R
\name{resid_data}
\alias{resid_data}
\title{Residualize data}
\usage{
resid_data(gamlssModel, df, og_data = NULL, rm_terms = NULL, zero_terms = NULL)
}
\arguments{
\item{gamlssModel}{gamlss model object}

\item{df}{dataframe to residualize. NOTE: gamlssModel will be refit to these data}

\item{og_data}{(optional) original dataframe on which model was fit, if differs from \code{df}}

\item{rm_terms}{list of term(s) whose effects will be residualized (set to mean/mode)}

\item{zero_terms}{list of NUMERIC term(s) whose effects will be removed completely (set to zero)}
}
\value{
dataframe with the outcome var of the gamlssModel residualized
}
\description{
Residualize data by removing terms' location effects as estimated by gamlss model
}
\details{
Works by running \code{predict.gamlss()} with type="terms" to estimate effects of specified terms/covariates on mu,
applying the inverse of the link function to convert into response scale, then subtracting from known y value.
Works with random effects, smooths, etc, but might have trouble correctly identifying terms
if they are not listed as they appear in \code{coefficients(gamlssModel)}.

IMPORTANT: Will not work if dataset has no variability (e.g. data with values simulated to hold constant).
}
