% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_funs.R
\name{sigma_predict}
\alias{sigma_predict}
\title{Predict sigma}
\usage{
sigma_predict(
  gamlssModel,
  sim_data_list,
  x_var,
  df = NULL,
  average_over = FALSE
)
}
\arguments{
\item{gamlssModel}{gamlss or gamlss2 model object}

\item{sim_data_list}{list of simulated dataframes returned by \code{\link[=sim_grid]{sim_grid()}}}

\item{x_var}{continuous predictor (e.g. 'age'), which \code{sim_data_list} varies over}

\item{df}{(optional) original dataframe from which new data will be simulated. Passing this can
fix some bugs in \code{\link[gamlss:predict.gamlss]{gamlss::predictAll()}}}

\item{average_over}{logical indicating whether to return sigma averaged across multiple
levels of a factor, with each level represented as a dataframe in \code{sim_data_list}.
Defaults to \code{FALSE}.}
}
\value{
list of dataframes containing predicted sigma across range of predictors
}
\description{
Calculates predicted sigma values across simulated data
}
\details{
This function takes a list of dataframes simulated with \code{\link[=sim_grid]{sim_grid()}} and calculates
the value of sigma (after link function is applied) as a way to visualize variability.
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris, family=BCCG)
sim_df <- sim_grid(iris, "Sepal.Length", "Species", iris_model)

#to average across levels of "Species"
sigma_predict(iris_model, sim_df, "Sepal.Length", average_over = TRUE)

}
