% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{simultaneous_pct_ci}
\alias{simultaneous_pct_ci}
\title{Simultaneous Confidence Intervals}
\usage{
simultaneous_pct_ci(df, x_var, col_name, interval)
}
\description{
Helper function for \code{gamlss_ci()}, \code{ci_type} = "simultaneous"
Implements simultaneous CIs along \code{x_var} as described in Gao, Konietschke, & Li, 2021 (doi: 10.3390/sym13071212).
}
