% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{sliding_window_ci}
\alias{sliding_window_ci}
\title{Sliding Window Confidence Intervals}
\usage{
sliding_window_ci(df, x_var, w, interval)
}
\description{
Helper function for \code{gamlss_ci()}, \code{ci_type} = "sliding"
}
