% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{un_log}
\alias{un_log}
\title{un-log}
\usage{
un_log(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
\code{un_log()} returns 10 raised to the power of a given input
}
\details{
Used to undo log(x, base=10) scaling
}
\examples{
x <- log(5, base=10)
un_log(x) #returns 5

}
