% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_funs.R
\name{wp.taki}
\alias{wp.taki}
\title{Worm Plot (Taki's Version)}
\usage{
wp.taki(
  object = NULL,
  xvar = NULL,
  resid = NULL,
  n.inter = 4,
  inter.breaks = NULL,
  xlim.worm = 3.5,
  ylim.worm = 12 * sqrt(n.inter/length(resid)),
  ...
)
}
\arguments{
\item{object}{gamlss model object}

\item{xvar}{vector containing values of predictor used for plotting model (requires \code{resid})}

\item{resid}{vector containing the residuals of the same model object (requires \code{xvar})}

\item{n.inter}{(optional) number of subsets with ~equal number of points to plot across the range of \code{xvar}. Defaults to 4.}

\item{xlim.worm}{control plot range(s)}

\item{ylim.worm}{control plot range(s)}
}
\value{
list with ggplot objects of worm plot(s) ($plot) and df of pts outside dotted CI ($outliers)
}
\description{
Output more robust worm plots for assessing gamlss fit in ggplot
}
\details{
This function fixes an inconsistency in the original package - that if do a worm plot with a formula
it gives the wrong answer compared to if you provide a vector of values to plot against. This is
rectified by adding a check that the partition of the covariate is actually providing a partition.
The percentage of points that fall outside the pointwise confidence intervals are shown in blue
(5\% indicates good fit). Written and contributed by the illustrious Taki Shinohara :)
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris, family=BCCG)

#get one worm plot by just passing the model object
wp.taki(iris_model)

#or get subplots across the range of a covariate
wp.taki(xvar=iris$Sepal.Length, resid=iris_model$residuals)
wp.taki(xvar=iris$Sepal.Length, resid=resid(iris_model))
wp.taki(xvar=iris$Sepal.Length, resid=resid(iris_model), n.inter=10)

}
