% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fan_helper_funs.R
\name{age_at_peak}
\alias{age_at_peak}
\title{Get Age of Peak}
\usage{
age_at_peak(cent_df, peak_from = NULL)
}
\arguments{
\item{cent_df}{dataframe output by \code{centile_predict()}}

\item{peak_from}{optional name of colum in \code{df} that's maximum will define the peak. Otherwise,
defaults to median centile ("cent_0.5")}
}
\value{
dataframe
}
\description{
Find age of peak of median or other specified centile estimates.
}
\details{
value (or peak change) for each factor level (using lapply). Defaults to finding the
age of the median centile's peak, but can find other specified centiles as well.
}
