% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{cent_cdf}
\alias{cent_cdf}
\title{Centile CDF}
\usage{
cent_cdf(gamlssModel, df, group = NULL, interval_var = NULL, ...)
}
\arguments{
\item{gamlssModel}{gamlss model object}

\item{df}{dataframe to assess}

\item{group}{(optional) name of grouping column}

\item{interval_var}{(optional) numeric variable along which to group outputs. Uses \link[ggplot2:cut_interval]{ggplot2::cut_interval}, which
requires additional args.}
}
\value{
dataframe with one row for every value of \code{group}
}
\description{
Return the probability of observations with predicted centiles that are < or = centile lines
estimated from a gamlss model.
}
\details{
Results can be grouped by any variable in the original dataframe. Inspired by output of \code{\link[gamlss:centiles]{gamlss::centiles()}}.
Calls \code{\link[=pred_og_centile]{pred_og_centile()}}.
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris)
cent_cdf(iris_model, iris)
cent_cdf(iris_model, iris, "Species")

}
