% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fans.R
\name{centile_fan_minimal}
\alias{centile_fan_minimal}
\title{Plot minimalist centile fan}
\usage{
centile_fan_minimal(
  gamlssModel,
  df,
  x_var,
  color_var = NULL,
  desiredCentiles = c(0.05, 0.5, 0.95),
  average_over = FALSE,
  sim_data_list = NULL,
  ...
)
}
\value{
ggplot object
}
\description{
Wrapper function for \code{\link[=make_centile_fan]{make_centile_fan()}} with different defaults.
}
