% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fans.R
\name{centile_fan_resid}
\alias{centile_fan_resid}
\title{Plot residualized centile fan}
\usage{
centile_fan_resid(
  gamlssModel,
  df,
  x_var,
  color_var = NULL,
  get_peaks = FALSE,
  x_axis = c("custom", "lifespan", "log_lifespan", "lifespan_fetal",
    "log_lifespan_fetal"),
  desiredCentiles = c(0.01, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99),
  average_over = FALSE,
  sim_data_list = NULL,
  show_points = TRUE,
  label_centiles = TRUE,
  remove_cent_effect,
  remove_point_effect = NULL,
  ...
)
}
\value{
ggplot object
}
\description{
Wrapper function to make it easier to plot centile fans in which the
effects of nuisance covariates are residualized out.
}
\details{
Calls \code{\link[=make_centile_fan]{make_centile_fan()}} but with different defaults. Is able to use covariates specified in
\code{remove_cent_effect} (which takes straightforward variable names) to find the and remove the same
covariates from individual data points, as necessary.
}
