% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{ci_diffs}
\alias{ci_diffs}
\title{Test Median Diffs across X}
\usage{
ci_diffs(ci_list)
}
\arguments{
\item{ci_list}{output of gamlssTools::gamlss_ci()}
}
\value{
dataframe
}
\description{
Uses bootstrapped CIs to find regions across \code{x_var} where the levels of a factor significantly differ
}
\details{
Only works for factors with 2 levels
}
