% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fan_helper_funs.R
\name{format_x_axis}
\alias{format_x_axis}
\title{Format X-Axis}
\usage{
format_x_axis(
  x_axis = c("custom", "lifespan", "log_lifespan", "lifespan_fetal",
    "log_lifespan_fetal"),
  x_values
)
}
\description{
Mostly internal/helper fun for formatting x-axis in centile fan plots
}
