% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{gamlss_try}
\alias{gamlss_try}
\title{gamlss try}
\usage{
gamlss_try(...)
}
\value{
gamlss model object
}
\description{
Try-catch fitting gamlss with various methods, return NULL if failed
}
\details{
Takes any \emph{named} gamlss model parameters. Tries quicker, default methods
(e.g. mu.step=1, method=RS()) before resorting to slower methods as necessary to fit. Returns NULL model
instead of giving errors, which is also useful when you need the script to continue
despite nonconvergence of some models.

NOTE: currently only fits gamlss models (not gamlss2). Also returns ugly call parameter in \code{summary()}.
}
\examples{
iris_model <- gamlss_try(formula = Sepal.Width ~ Sepal.Length + Petal.Width + Species, sigma.formula = ~ Sepal.Length, data=iris, family=NO)

#make sure you name any parameters you pass! unnamed formula param will fail:
\dontrun{
iris_model <- gamlss_try(Sepal.Width ~ Sepal.Length + Petal.Width + Species, sigma.formula = ~ Sepal.Length, data=iris, family=NO)
}
}
