% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fan_helper_funs.R
\name{get_derivatives}
\alias{get_derivatives}
\title{Get Centile Derivative}
\usage{
get_derivatives(cent_df)
}
\arguments{
\item{cent_df}{dataframe output by \code{centile_predict()}}
}
\value{
dataframe
}
\description{
Find first-order derivative of any centile lines in dataframe
}
\details{
Takes output of \code{centile_predict()} and finds derivatives of centile fans at each
factor level (using lapply).
}
