% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{list_predictors}
\alias{list_predictors}
\title{List all predictors}
\usage{
list_predictors(gamlssModel)
}
\arguments{
\item{gamlssModel}{gamlss model object}
}
\value{
dataframe with outputs of \code{drop1()} for each moment and term
}
\description{
Lists every covariate in any moment of a gamlss model.
}
\details{
Does not distinguish smooth, fixed, or random effects.
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris)
list_predictors(iris_model)

}
