% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fans.R
\name{plot_centile_cis}
\alias{plot_centile_cis}
\title{Plot Centiles with CIs}
\usage{
plot_centile_cis(
  gamlssModel,
  df,
  x_var,
  color_var,
  desiredCentiles = c(0.5),
  interval = 0.95,
  B = 100,
  sim_data_list = NULL,
  type = c("resample", "bayes", "LOSO"),
  stratify = FALSE,
  boot_group_var = NULL,
  special_term = NULL,
  boot_list = NULL,
  average_over = FALSE,
  ci_type = "simultaneous",
  ...
)
}
\arguments{
\item{gamlssModel}{gamlss model object}

\item{df}{dataframe used to fit the gamlss model}

\item{x_var}{continuous predictor (e.g. 'age') that will be plotted on the x axis}

\item{color_var}{(optional) categorical predictor (e.g. 'sex') that will be used to determine the color of
points/centile lines. Alternatively, you can average over each level of this variable
to return a single set of centile lines (see \code{average_over}).}

\item{desiredCentiles}{list of percentiles as values between 0 and 1 that will be
calculated and returned. Defaults to c(0.01, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99),
which returns the 1st percentile, 5th percentile, 10th percentile, etc.}

\item{interval}{size of confidence interval to calculate. Defaults to 0.95, or 95\%}

\item{B}{number of samples/models to bootstrap. Defaults to 100. if \code{type = "LOSO"}, B will be updated to
the number of unique values of \code{group_var}}

\item{sim_data_list}{(optional) output of \code{sim_data()}}

\item{type}{which type of bootstrapping to perform. \code{resample} performs traditional bootstrapping (resample with replacement)
across all groups; alternatively, it may be combined with \code{stratify=TRUE} and \code{group_var} args below to bootstrap
while maintaining each group's (e.g study's) n. \code{bayes} keeps the original dataframe but randomizes each observation's
weight. \code{LOSO} drops an entire subset from the sample (indicated by \code{group_var}) with each bootstrap.}

\item{stratify}{logical. with \code{type=resample} will bootstrap within each level of \code{group_var}.}

\item{boot_group_var}{categorical/factor variable that resampling will be stratified within (when \code{type=resample})
or that one level will be dropped from in each bootstraped sample (when \code{type=LOSO}). Can also be a list, allowing
stratification within multiple groups e.g. \code{group_var=c(sex, study)}}

\item{special_term}{optional, passed to gamlssTools::sim_data()}

\item{boot_list}{(optional) output of gamlssTools::bootstrap_gamlss()}

\item{average_over}{logical indicating whether to average predicted centiles across each level of \code{color_var}.
Defaults to \code{FALSE}, which will plot a different colored centile fan for each level of \code{color_var}.}
}
\value{
ggplot object
}
\description{
Plot centile fan with confidence intervals on 50th centile
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris, family=BCCG)
plot_centile_cis(iris_model, iris, "Sepal.Length", "Species", stratify=TRUE, boot_group_var="Species")

}
