% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fans.R
\name{plot_centile_deriv}
\alias{plot_centile_deriv}
\title{Plot Derivatives}
\usage{
plot_centile_deriv(
  gamlssModel,
  df,
  x_var,
  color_var,
  desiredCentiles = c(0.5),
  sim_data_list = NULL,
  ...
)
}
\value{
ggplot object
}
\description{
Plot 1st derivative of centile line
}
\details{
Wrapper function for \code{\link[=make_centile_fan]{make_centile_fan()}} with different defaults. Defaults
to plotting only derivative of median line, but can be updated
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris, family=BCCG)
plot_centile_deriv(iris_model, iris, "Sepal.Length", "Species")

}
