% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{pred_og_centile}
\alias{pred_og_centile}
\title{Predict original centiles}
\usage{
pred_og_centile(gamlssModel, og.data, get.zscores = FALSE)
}
\arguments{
\item{gamlssModel}{gamlss model object}

\item{og.data}{dataframe used to fit \code{gamlssModel}}

\item{get.zscores}{logical indicating whether to calculate and return z-scores from centiles}
}
\value{
either vector listing centiles for every datapoint OR dataframe with centiles and z-scores
}
\description{
Returns the centile and/or z-score values for the original datapoints used to fit a gamlss model
}
\details{
Based on Jenna's function \href{https://github.com/jmschabdach/mpr_analysis/blob/70466ccc5f8f91949b22745c227017bf47ab825c/r/lib_mpr_analysis.r#L67}{calculatePhenotypeCentile()}
and \code{\link[gamlss:z.scores]{gamlss::z.scores()}}. Currently only supports z-score calculations for BCCG and NO families of distributions, can add others
as appropriate.
}
\examples{
iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris)
list_predictors(iris_model)

}
