% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fan_helper_funs.R
\name{remove_effects}
\alias{remove_effects}
\alias{resid_data}
\title{Remove estimated effects from data}
\usage{
remove_effects(gamlssModel, data, ref_data = NULL, terms)

resid_data(gamlssModel, df, og_data = NULL, rm_terms)
}
\arguments{
\item{gamlssModel}{gamlss model object}

\item{data}{dataframe to residualize. NOTE: gamlssModel will be refit to these data}

\item{ref_data}{(optional) original dataframe on which model was fit.}

\item{terms}{list of term(s) whose effects will be residualized (removed).}
}
\value{
dataframe with the outcome var of the gamlssModel residualized
}
\description{
Residualize data by removing terms' location effects as estimated by a gamlss model
}
\details{
Works by running \code{predict.gamlss()} with type="terms" to estimate effects of specified terms/covariates on mu,
applying the inverse of the link function to convert into response scale, then subtracting from known y value.

IMPORTANT: Will not work if dataset has no variability (e.g. data with values simulated to hold constant). To remove
estimated covariate effects from simulated data, use \code{\link[=pred_residualized]{pred_residualized()}}.

Works with random effects, smooths, etc, but might have trouble correctly identifying terms
if they are not listed as they appear in \code{coefficients(gamlssModel)}.

By default (\code{ref_data=NULL}) the model's location effects are reconstructed WITHOUT the original data
whenever possible; supplying \code{ref_data} forces prediction via
\code{\link[gamlss:predict.gamlss]{gamlss::predictAll()}} with those data, which is also done automatically when the model
contains a non-reconstructable smoother (\code{cs()}, \code{ps()}, \code{ga()}, \code{s()}).

\code{resid_data()} is a deprecated alias for \code{remove_effects()}. Its \code{df}, \code{og_data}
and \code{rm_terms} arguments map onto \code{data}, \code{ref_data} and \code{terms} respectively.
}
