% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{safe_gamlss}
\alias{safe_gamlss}
\title{safe gamlss}
\usage{
safe_gamlss(...)
}
\value{
gamlss model object
}
\description{
gamlss() with more error handling
}
\details{
Fits model using \code{\link[gamlss:gamlss]{gamlss::gamlss()}} and throws an error if model fails to converge or is null

NOTE: currently only fits gamlss models (not gamlss2). Also returns ugly call parameter in \code{\link[gamlss:summary]{gamlss::summary()}}.
}
\examples{
iris_model <- safe_gamlss(formula = Sepal.Width ~ Sepal.Length + Petal.Width + Species, sigma.formula = ~ Sepal.Length, data=iris, family=NO)

}
