% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile_fan_helper_funs.R
\name{sim_grid}
\alias{sim_grid}
\alias{sim_data}
\title{Simulate a covariate grid for plotting GAMLSS}
\usage{
sim_grid(
  ref_data,
  x_var,
  factor_var = NULL,
  gamlssModel = NULL,
  special_term = NULL,
  x_range = NULL
)

sim_data(
  df,
  x_var,
  factor_var = NULL,
  gamlssModel = NULL,
  special_term = NULL,
  x_range = NULL
)
}
\arguments{
\item{ref_data}{original dataframe from which grid will be simulated}

\item{x_var}{continuous variable whose value will be simulated across it's full range,
as determined from the \code{ref_data} parameter}

\item{factor_var}{(optional) categorical variable that will be simulated at every level}

\item{gamlssModel}{(optional) gamlss model object that will be used to remove any extra columns
in \code{ref_data} such that only the model's covariates are simulated across}

\item{special_term}{formula (or character vector of formulas) defining any terms that should be calculated separately
(e.g. interaction terms). When passing multiple, list them in dependency order — each is applied in sequence.
Can also be used to set term to zero (i.e. \code{special_term="Sepal.Length = 0"}).}

\item{x_range}{x-values for which values are estimated (e.g. seq(0, 10)). Default is 500 datapoints across range of x axis.}
}
\value{
list of dataframes of simulated data, one for each level of \code{factor_var}
}
\description{
\code{sim_grid()} creates a data grid from which you can cleanly plot centiles.
}
\details{
This function takes a dataset and transforms it such that covariate(s) of interest are
allowed to vary across their full range of values, while all other covariates are held
constant at their reference (mode or mean) value. This enables clean visualization of
centiles across, for example, age, while holding freesurfer version constant. Can be
used on its own or as a subfunction of \code{\link[=make_centile_fan]{make_centile_fan()}}.

\code{sim_data()} is a deprecated alias for \code{sim_grid()}.
}
\examples{
sim_grid(iris, "Sepal.Length", "Species")

iris_model <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species, sigma.formula = ~ Sepal.Length, data=iris)
sim_grid(iris, "Sepal.Length", "Species", iris_model)

# add interaction term for dummy-coded species
iris2 <- iris \%>\%
  mutate(Species=as.numeric(Species)) \%>\%
  mutate(SL_int = Sepal.Length * Species)

iris_model2 <- gamlss(formula = Sepal.Width ~ Sepal.Length + Species + SL_int, sigma.formula = ~ Sepal.Length, data=iris2)
sim_grid(iris2, "Sepal.Length", "Species", iris_model2, special_term="SL_int = Sepal.Length * Species")

}
