% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_helper_funs.R
\name{trajectory_diff}
\alias{trajectory_diff}
\title{Get Diffs in Trajectories}
\usage{
trajectory_diff(
  gamlssModel,
  df,
  x_var,
  factor_var,
  sim_data_list = NULL,
  moment = c("mu", "sigma"),
  factor_var_levels = NULL,
  ...
)
}
\arguments{
\item{gamlssModel}{gamlss model object}

\item{df}{dataframe model was originally fit on}

\item{x_var}{continuous predictor (e.g. 'age'), which \code{sim_data_list} varies over}

\item{factor_var}{categorical variable to compare levels within.}

\item{sim_data_list}{list of simulated dataframes returned by \code{sim_grid()}}

\item{moment}{what moment to get differences for. \code{mu} calculates differences in 50th centile,
\code{sigma} calculates differences in predicted value of sigma (with link-function applied)}

\item{factor_var_levels}{(optional) specify the order for factor levels. E.g., \code{factor_var_levels = c("A", "B")}
would calculate the difference A - B.}
}
\value{
dataframe
}
\description{
Calculate differences in 50th centile or sigma trajectories between 2 factor levels
}
\details{
To test significance, see \code{\link[=ci_diffs]{ci_diffs()}}
}
